requirements.txt
datasets>=2.2.1
huggingface-hub>=0.4.0
matplotlib>=3.3.4
nltk>=3.6.7
notebook>=6.4.10
numpy>=1.19.5
pandas>=1.1.5
pykeops==2.1.1
scikit-learn>=0.24.2
tokenizers>=0.12.1
torch>=1.10.2
torchinfo>=1.5.4
tqdm>=4.64.0
transformers>=4.18.0
umap-learn>=0.5.3